% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_bgej.R
\docType{data}
\name{bgej}
\alias{bgej}
\title{bgej (DATA) Summary Indexes for Census blockgroups}
\description{
bgej is a table of all blockgroups, with the raw scores of the Summary Indexes
and Supplemental Summary Indexes for all the environmental indicators.
}
\details{
This file is not stored in the package.

For documentation on the residential population and environmental data and indicators,
see \href{https://web.archive.org/web/20250118193121/https://www.epa.gov/ejscreen/understanding-ejscreen-results}{EJSCREEN documentation}.

The column names are these:

\if{html}{\out{<div class="sourceCode">}}\preformatted{c('bgfips', 'bgid', 'ST', 'pop',
names_ej,
names_ej_supp,
names_ej_state,
names_ej_supp_state
)
}\if{html}{\out{</div>}}
}
