% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_PACKAGE_DEV.R
\name{find_in_files}
\alias{find_in_files}
\title{utility to do global search/find in full text of the files in a folder, like source code files or unit tests}
\usage{
find_in_files(
  pattern,
  path = "./tests/testthat",
  filename_pattern = "\\\\.R$|\\\\.r$",
  ignorecomments = FALSE,
  ignore.case = TRUE,
  value = TRUE,
  whole_line = TRUE,
  quiet = TRUE
)
}
\arguments{
\item{pattern}{regular expression to look for}

\item{path}{can change it to e.g., "./R"}

\item{filename_pattern}{query regex on file names, default is R code files}

\item{ignorecomments}{omit hits from commented out lines}

\item{ignore.case}{as in grep}

\item{value}{logical as in \code{\link[=grep]{grep()}} if TRUE returns matching text;
if FALSE, returns logical vectors like \code{\link[=grepl]{grepl()}}}

\item{whole_line}{set it to FALSE to see only the matching fragments
vs entire line of text that has a match in it}

\item{quiet}{whether to print results or just invisibly return}
}
\value{
list of named vectors,
where names are file paths with hits, elements are vectors of text with hits
}
\description{
utility to do global search/find in full text of the files in a folder, like source code files or unit tests
}
\examples{
EJAM:::find_in_files("[^_]logo_....",    path = "./R", whole_line = FALSE, quiet = F)
EJAM:::find_in_files("report_logo.....", path = "./R", whole_line = FALSE, quiet = F)
EJAM:::find_in_files("app_logo......",   path = "./R", whole_line = FALSE, quiet = F)

EJAM:::find_in_files("latlon_from_.{18}", quiet = FALSE, whole_line = F)
EJAM:::find_in_files("latlon_from_s.{9}", quiet = FALSE, whole_line = F)
EJAM:::find_in_files("latlon_from_mact.{9}", quiet = FALSE, whole_line = F)

}
\keyword{internal}
