% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIPS_FUNCTIONS.R
\name{fips2statefips}
\alias{fips2statefips}
\alias{fips2state_fips}
\title{FIPS - Get FIPS codes of the States CONTAINING the given census units (of any type)}
\usage{
fips2statefips(fips)
}
\arguments{
\item{fips}{vector of FIPS}
}
\value{
vector of State FIPS 2 characters each
}
\description{
FIPS - Get FIPS codes of the States CONTAINING the given census units (of any type)
}
\details{
Tells you which State contains each County (or tract or blockgroup or block)
}
