% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIPS_FUNCTIONS.R
\name{fips2statename}
\alias{fips2statename}
\title{FIPS - Get names of the States CONTAINING the given census units (of any type)}
\usage{
fips2statename(fips)
}
\arguments{
\item{fips}{vector of FIPS}
}
\value{
vector of state names
}
\description{
FIPS - Get names of the States CONTAINING the given census units (of any type)
}
