% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fips_xyz_from_latlon.R
\name{fips_county_from_latlon}
\alias{fips_county_from_latlon}
\title{Find what county is where each point is located}
\usage{
fips_county_from_latlon(sitepoints = NULL, lat = NULL, lon = NULL)
}
\arguments{
\item{sitepoints}{data.frame with lat,lon columns}

\item{lat}{latitudes vector if sitepoints not provided}

\item{lon}{longitudes vector if sitepoints not provided}
}
\value{
just vector of fips
}
\description{
Find what county is where each point is located
}
\keyword{internal}
