% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fips_xyz_from_latlon.R
\name{fips_state_from_latlon}
\alias{fips_state_from_latlon}
\title{Find what state is where each point is located}
\usage{
fips_state_from_latlon(sitepoints = NULL, lat = NULL, lon = NULL)
}
\arguments{
\item{sitepoints}{data.frame with lat,lon columns}

\item{lat}{latitudes vector if sitepoints not provided}

\item{lon}{longitudes vector if sitepoints not provided}
}
\value{
just vector of fips, unlike \code{\link[=state_from_latlon]{state_from_latlon()}}
}
\description{
Find what state is where each point is located
}
\seealso{
\code{\link[=state_from_latlon]{state_from_latlon()}}
}
\keyword{internal}
