% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_locate_by_id.R
\name{locate_by_id}
\alias{locate_by_id}
\title{DRAFT - query FRS API to find latest info on EPA facilities by registry ID or program ID}
\usage{
locate_by_id(idx, type = "frs", simple = FALSE, ...)
}
\arguments{
\item{idx}{vector of one or more character strings that must be
registry IDs (default) or program IDs}

\item{type}{either frs (default) or program.
frs means all are registry_id and
program means all are pgm_sys_id}

\item{simple}{use simpler method}

\item{...}{passed through to \code{\link[=locate_by_id1]{locate_by_id1()}}}
}
\value{
data.frame with one row per queried id, columns as returned by API
but lat lon instead of Latitude83 Longitude83
}
\description{
FUNCTION IS NOT FULLY WORKING/TESTED
}
\details{
May want to use \code{\link[=latlon_from_programid]{latlon_from_programid()}} and \code{\link[=latlon_from_regid]{latlon_from_regid()}} instead.

Uses the Facility Registry Service (FRS) API to find a site
by registry ID or program ID.
This uses an API to find sites, but it is faster to look in a table
if that FRS dataset is already loaded in an app, for example.

see \code{\link[=url_by_id]{url_by_id()}} for more details
}
\examples{
  \donttest{

  ids <- testinput_program_sys_id
  ids <- paste0(ids$program, ids$pgm_sys_id)
  # ids <- c('ILR000128264','600039382','TXR1592DZ','TSCA8851',
  #    'CT0000000900908716', 'CEDRI10043548',  'CO0000000812305826')
  # ids <- c('ILR000128264','600039382')
  locate_by_id(ids, type = 'program')   # stopped working   ***********
}

}
\seealso{
?latlon_from_programid() and ?latlon_from_regid() and url_by_id() and the obsolete locate_by_id()
}
\keyword{internal}
