% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_metadata_add.R
\name{metadata_check_print}
\alias{metadata_check_print}
\title{helper function in updating the package metadata, used by scripts in /data-raw/
prints to console info about some missing metadata}
\usage{
metadata_check_print(...)
}
\value{
same as \code{\link[=metadata_check]{metadata_check()}}, invisibly
}
\description{
helper function in updating the package metadata, used by scripts in /data-raw/
prints to console info about some missing metadata
}
\seealso{
\code{\link[=metadata_check_print]{metadata_check_print()}} \code{\link[=metadata_check]{metadata_check()}} \code{\link[=metadata_add]{metadata_add()}} \code{\link[=metadata_update_attr]{metadata_update_attr()}}
}
\keyword{internal}
