% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popup_from_ejscreen.R
\name{popup_from_ejscreen}
\alias{popup_from_ejscreen}
\title{Map popups - Create the popup text for maps of EJ results}
\usage{
popup_from_ejscreen(
  out,
  linkcolnames = sapply(EJAM:::global_or_param("default_reports"), function(x) x$header),
  verbose = FALSE
)
}
\arguments{
\item{out}{like ejamit()$results_bysite, but also it can be full list from ejamit().
The table of raw data in data.frame form, with results of EJ analysis.}

\item{linkcolnames}{Vector of column names in the table that have links to URLs like reports about single sites}

\item{verbose}{TRUE or FALSE, can see more details reported when function is used.}
}
\value{
HTML ready to be used for map popups
}
\description{
This creates the HTML text that appears in map popups.
}
\details{
Popup shows up in a window when you
click on a site on the map, when viewing the results of EJ analysis of
each site.

THIS IS CURRENTLY HARD CODED TO USE EJScreen VARIABLE NAMES.

It provides raw scores (but not for summary indexes) and US and State percentiles if available:
\itemize{
\item some site id info fields if found
\item latitude, longitude, and size of area around site
\item Indicators like population count, Residential Population Demog. Indicator, etc.
\item Environmental Indicators
\item Summary Indexes
\item web link(s) to map or report
}
}
\keyword{internal}
