% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_testinput_fips_mix.R
\docType{data}
\name{testinput_fips_mix}
\alias{testinput_fips_mix}
\title{testinput_fips_mix dataset}
\format{
An object of class \code{character} of length 6.
}
\usage{
testinput_fips_mix
}
\description{
Census FIPS codes vector as example, e.g., mapfast(shapes_from_fips(testinput_fips_mix))
}
\keyword{datasets}
