% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_testinput_regid.R
\docType{data}
\name{testinput_regid}
\alias{testinput_regid}
\title{test data, EPA Facility Registry ID numbers to try using}
\format{
An object of class \code{numeric} of length 8.
}
\usage{
testinput_regid
}
\description{
test data, EPA Facility Registry ID numbers to try using
}
\details{
Just for convenience, installed with the package
}
\keyword{datasets}
