% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/URL_FUNCTIONS_part2.R
\name{url_countyhealth}
\alias{url_countyhealth}
\title{URL functions - Get URLs of useful report(s) on Counties containing the given fips, from countyhealthrankings.org}
\usage{
url_countyhealth(
  fips = NULL,
  year = 2025,
  sitepoints = NULL,
  shapefile = NULL,
  as_html = FALSE,
  linktext = "County",
  ifna = "https://www.countyhealthrankings.org",
  baseurl = "https://www.countyhealthrankings.org/health-data/",
  ...
)
}
\arguments{
\item{fips}{vector of fips codes of counties, 5 characters each, like "10003",
or other fips codes like blockgroups in which case it returns info for parent counties}

\item{year}{e.g., 2025}

\item{sitepoints}{if provided and fips is NULL, gets county fips from lat,lon columns of sitepoints}

\item{shapefile}{if provided and fips is NULL, gets county fips from lat,lon of polygon centroid}

\item{as_html}{Whether to return as just the urls or as html hyperlinks to use in a DT::datatable() for example}

\item{linktext}{used as text for hyperlinks, if supplied and as_html=TRUE}

\item{ifna}{URL shown for missing, NA, NULL, bad input values}

\item{baseurl}{do not change unless endpoint actually changed}

\item{...}{unused}
}
\value{
vector of URLs
}
\description{
URL functions - Get URLs of useful report(s) on Counties containing the given fips, from countyhealthrankings.org
}
\examples{
 url_countyhealth(fips_counties_from_state_abbrev("DE"))
 # browseURL(url_countyhealth(fips_counties_from_state_abbrev("DE"))[1])

}
