% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/URL_FUNCTIONS_part3.R
\name{url_ejscreen_report}
\alias{url_ejscreen_report}
\title{URL functions - API WENT DOWN? Get URLs of EJSCREEN reports}
\usage{
url_ejscreen_report(
  sitepoints = NULL,
  lat = "",
  lon = "",
  radius = "",
  mobile = FALSE,
  areatype = "",
  areaid = "",
  namestr = "",
  shapefile = NULL,
  fips = NULL,
  wkid = 4326,
  unit = 9035,
  f = "report",
  interactiveprompt = FALSE,
  as_html = FALSE,
  linktext = "Report",
  ifna = "https://ejanalysis.com",
  baseurl = NULL,
  ...
)
}
\arguments{
\item{sitepoints}{data.frame with colnames lat, lon (or lat, lon parameters can be provided separately)}

\item{lat}{one or more latitudes (or a table with lat, lon columns, or filepath with that, or omit to interactively select file)}

\item{lon}{one or more longitudes (or omitted -- see lat parameter details)}

\item{radius}{miles radius}

\item{mobile}{If TRUE, provides URL for the mobile browser version, not desktop version}

\item{areatype}{passed as areatype= in API, inferred if not provided but areaid is provided}

\item{areaid}{fips codes if used,  passed as areaid=  in API, can be FIPS for blockgroups, tracts, counties.
Not zip code! For example, "10001" will report on that county, not that zip code.
Note that nearly half of all county fips codes are impossible to distinguish from
5-digit zipcodes because the same numbers are used for both purposes.

For county FIPS 10001, use \code{url_ejscreen_report(areaid =  '10001')}

For zipcode 10001, you cannot use \code{url_ejscreen_report()} for zip codes
since the API does not support zip codes.
You can still at least try \code{url_ejscreenmap(wherestr =  '10001')}}

\item{namestr}{The character string text to show on the report as the name of the place}

\item{shapefile}{not implemented for old EJSCREEN API, should work for newer EJAM-API}

\item{fips}{not implemented for old EJSCREEN API, should work for newer EJAM-API}

\item{wkid}{default is 4326 -WGS84 - World Geodetic System 1984, used in GPS - see (https://epsg.io/4326)}

\item{unit}{default is 9035 which means miles; for kilometers use 9036}

\item{f}{can be "report" or "pjson" or "json"}

\item{interactiveprompt}{passed to \code{\link[=sitepoints_from_any]{sitepoints_from_any()}}}

\item{as_html}{Whether to return as just the urls or as html hyperlinks to use in a DT::datatable() for example}

\item{linktext}{used as text for hyperlinks, if supplied and as_html=TRUE}
}
\value{
URL(s)
}
\description{
Old function for pre-2025 EPA EJSCREEN report (see newer url_ejamapi() now)
}
\details{
If old EJSCREEN api is down, tries to use newer \code{\link[=url_ejamapi]{url_ejamapi()}}

For info on the old EJSCREEN API pre-Feb-2025,
see \code{\link[=ejscreenRESTbroker]{ejscreenRESTbroker()}} and \code{\link[=ejscreenapi1]{ejscreenapi1()}}

and \href{https://web.archive.org/web/20250201025755/https://ejscreen.epa.gov/mapper/EJAPIinstructions.pdf}{archived EJAPIinstructions.pdf}

and \href{https://web.archive.org/web/20250119121857/https://ejscreen.epa.gov/mapper/ejscreenapi1.html}{archived ejscreenapi1 page}
}
\seealso{
\code{\link[=url_ejscreen_report]{url_ejscreen_report()}}  \code{\link[=url_ejscreenmap]{url_ejscreenmap()}}
\code{\link[=url_echo_facility]{url_echo_facility()}} \code{\link[=url_frs_facility]{url_frs_facility()}}  \code{\link[=url_enviromapper]{url_enviromapper()}}
}
\keyword{internal}
