% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/URL_FUNCTIONS_part2.R
\name{url_enviromapper}
\alias{url_enviromapper}
\title{Get URLs of EnviroMapper reports}
\usage{
url_enviromapper(
  sitepoints = NULL,
  lon = NULL,
  lat = NULL,
  shapefile = NULL,
  fips = NULL,
  zoom = 13,
  as_html = FALSE,
  linktext = "EnviroMapper",
  ifna = "https://geopub.epa.gov/myem/efmap/",
  baseurl = "https://geopub.epa.gov/myem/efmap/index.html?ve=",
  ...
)
}
\arguments{
\item{sitepoints}{data.frame with colnames lat, lon (or lat, lon parameters can be provided separately)}

\item{lat, lon}{ignored if sitepoints provided, can be used otherwise, if shapefile and fips not used}

\item{shapefile}{if provided function uses centroids of polygons for lat lon}

\item{fips}{ignored}

\item{zoom}{initial map zoom extent}

\item{as_html}{Whether to return as just the urls or as html hyperlinks to use in a DT::datatable() for example}

\item{linktext}{used as text for hyperlinks, if supplied and as_html=TRUE}

\item{ifna}{URL shown for missing, NA, NULL, bad input values}

\item{baseurl}{do not change unless endpoint actually changed}

\item{...}{unused}
}
\value{
URL of one webpage (that launches the mapping tool)
}
\description{
Get URL(s) for EnviroMapper web-based tool, to open map at specified point location(s)
}
\details{
EnviroMapper lets you view EPA-regulated facilities and other information on a map, given the lat,lon
}
\examples{
x = url_enviromapper(testpoints_10[1,])
\dontrun{
 browseURL(x)
 browseURL(url_enviromapper(lat = 38.895237, lon = -77.029145, zoom = 17))
}

}
\seealso{
\code{\link[=url_ejamapi]{url_ejamapi()}}  \code{\link[=url_ejscreenmap]{url_ejscreenmap()}}
\code{\link[=url_echo_facility]{url_echo_facility()}} \code{\link[=url_frs_facility]{url_frs_facility()}}  \code{\link[=url_enviromapper]{url_enviromapper()}}
}
