% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/URL_FUNCTIONS_part2.R
\name{url_naics.com}
\alias{url_naics.com}
\title{URL functions - url_naics.com - Get URL for page with info about industry sectors by text query term}
\usage{
url_naics.com(
  query = "",
  as_html = FALSE,
  linktext = query,
  ifna = "https://www.naics.com",
  baseurl = "https://www.naics.com/code-search/?trms=",
  ...
)
}
\arguments{
\item{query}{string query term like "gasoline" or "copper smelting"}

\item{as_html}{Whether to return as just the urls or as html hyperlinks to use in a DT::datatable() for example}

\item{linktext}{used as text for hyperlinks, if supplied and as_html=TRUE}

\item{ifna}{URL shown for missing, NA, NULL, bad input values}

\item{baseurl}{do not change unless endpoint actually changed}

\item{...}{unused}
}
\value{
URL as string
}
\description{
URL functions - url_naics.com - Get URL for page with info about industry sectors by text query term
}
\details{
See (https://naics.com) for more information on NAICS codes.

Unlike url_xyz() functions, which provide a unique link for each site,
this url_ function provides just a link for a whole industry or set of industries based on a query,
so it is not meant to be used in a column of site by site results the way the other url_xyz() functions are.
}
