% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/URL_FUNCTIONS_part1.R
\name{url_xl_style}
\alias{url_xl_style}
\title{utility to prep URLs for being written to Excel}
\usage{
url_xl_style(urls, urltext = urls)
}
\arguments{
\item{urls}{vector of urls such as from \code{\link[=url_ejamapi]{url_ejamapi()}}}

\item{urltext}{The text to appear in Excel cells instead of just the URL showing}
}
\description{
utility to prep URLs for being written to Excel
}
\details{
See table_xls_format()

Works best if using \code{\link[openxlsx:writeData]{openxlsx::writeData()}} not \code{\link[openxlsx:write.xlsx]{openxlsx::write.xlsx()}}

To write this column of urls to a worksheet:

\if{html}{\out{<div class="sourceCode">}}\preformatted{lat <- c(30.977402, 32.515813); lon = c(-83.368997, -86.377325)
radius <- 1
urls <- url_ejscreenmap(lat=lat, lon=lon, radius=radius)

urlx <- EJAM:::url_xl_style(urls, urltext = paste0("Report ", 1:2))

wb <- openxlsx::createWorkbook()
openxlsx::addWorksheet(wb, sheetName = 'tab1')
openxlsx::writeData(wb, sheet = 1, x = urlx, startCol = 1, startRow = 2)
openxlsx::saveWorkbook(wb, file = '~/test1.xlsx', overwrite = TRUE)

# using just [openxlsx::write.xlsx()] is simpler but ignores the urltext param:
openxlsx::write.xlsx(data.frame(lat = lat, lon = lon, urlx), file = 'test2.xlsx')
}\if{html}{\out{</div>}}
}
\keyword{internal}
