% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_interactive_table.R
\name{create_interactive_table}
\alias{create_interactive_table}
\title{creates interactive html version of site by site table in app}
\usage{
create_interactive_table(
  out,
  reports = EJAM:::global_or_param("default_reports"),
  sitereport_download_buttons_colname = "Download EJAM Report",
  sitereport_download_buttons_show = TRUE,
  columns_used = NULL
)
}
\arguments{
\item{out}{list of tables like data_processed in app_server, similar to output of ejamit()}

\item{reports}{ignored for now - info about which URLs/links/reports columns to include among those already in out}

\item{sitereport_download_buttons_colname}{header for column to create with buttons to download 1-site reports in shiny app}

\item{sitereport_download_buttons_show}{if TRUE, add column near first with buttons to allow download of 1-site html summary report}

\item{columns_used}{if specified in server based on defaults or inputs, these are a subset of
colnames from ejamit()$results_bysite  to show in site-by-site interactive table}
}
\description{
Builds site by site table after an analysis in EJAM app. Pulls in uploaded and analyzed data to create table
}
\seealso{
\code{\link[=ejam2tableviewer]{ejam2tableviewer()}}
}
\keyword{internal}
