% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIPS_FUNCTIONS.R
\name{fips2countyfips}
\alias{fips2countyfips}
\title{Get PARENT COUNTY FIPS of each fips (the County that contains the fips)}
\usage{
fips2countyfips(fips)
}
\arguments{
\item{fips}{vector of FIPS codes where fipstype(fips) is among
block, blockgroup, tract, city, county, state
but where it is a state this will return NA}
}
\value{
vector of fips as long as input
}
\description{
Get PARENT COUNTY FIPS of each fips (the County that contains the fips)
}
\examples{
fips2countyfips(testinput_fips_blockgroups[1])

}
